BACK-END/study_language/requirements.txt
Django==4.2.23
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
gunicorn==21.2.0
django-environ==0.11.2
dj-database-url
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
